Refactor: Move message package to common.
Trying to clean up the top-level directory a little bit.

PiperOrigin-RevId: 291098651
