+++ probes/udplistener/udplistener.go
@@ -45,8 +45,8 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/google/cloudprober/common/message"
 	"github.com/google/cloudprober/logger"
-	"github.com/google/cloudprober/message"
 	"github.com/google/cloudprober/metrics"
 	"github.com/google/cloudprober/probes/common/statskeeper"
 	"github.com/google/cloudprober/probes/options"
